requirements.txt
# DEX MCP Server Dependencies

# Core MCP framework
fastmcp>=0.1.0

# HTTP client for API requests
aiohttp>=3.8.0

# Faster event loop (Unix only)
uvloop>=0.17.0; sys_platform != "win32"

# Optional development dependencies (uncomment if needed)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=22.0.0
# isort>=5.10.0
# mypy>=1.0.0 